requirements.txt
hexbytes==0.3.0
langchain==0.0.207
protobuf==4.23.3
pydantic==1.10.7
PyQt5==5.15.9
python-dotenv==1.0.0
SpeechRecognition==3.10.0
web3==6.3.0
